tweaks

+++ php/clearCart.php
@@ -1,6 +1,7 @@
 <?php
-
-$query = "DELETE FROM cart";
+session_start();
+$user_id = $_SESSION['user_id'];
+$query = "DELETE FROM cart WHERE client_id = $user_id";
 
 include "oopattempt.php";
 $connection = new Connection("localhost", "root", "", "boblustrations");
@@ -8,6 +9,6 @@ $connection->init_conn();
 
 $connection->conn->query($query);  
 $conn = null;
-
+header("location: ../cart.php");
 
 ?>
